2014 03 19 13:19
Alteraçãoes de metodo nas classes!

+++ Alg3_M_E1/src/Arquivos/Ficheiro.java
@@ -1,5 +1,6 @@
 package Arquivos;
 
+import java.awt.PageAttributes;
 import java.io.BufferedReader;
 import java.io.File;
 import java.io.FileInputStream;
@@ -25,46 +26,49 @@ public class Ficheiro {
 	 * 
 	 * @param nomeAquivo
 	 */
-	public void criaNovoArquivo(final String nomeAquivo){
+	public void criaNovo(final String nomeAquivo){
 		try {
 			FileWriter c = new FileWriter(new File(nomeAquivo),true);
-			i.msg("Novo arquivo criado: "+c);
+			i.msgb("Novo arquivo criado: "+c);
 			c.close();
 		} catch (Exception e) {
-			i.msg("Não foi possivél criar um novo arquivo!");
+			i.msgb("Não foi possivél criar um novo arquivo!");
 		}		
 	}
 	
-	/**
-	 *   
+	/** Metodo de remoção de arquivo com verificação de exintencia do arquivo.
 	 * @param nomeArquivo
 	 */
 	public void removeArquivo(String nomeArquivo){
 		String nome = nomeArquivo; 
 		File f = new File(nome); 
-		f.delete();
-		i.msg("O aquivo " +nome +" foi removido!");
+		if (f.exists()){ 
+			i.msgb("???" +f.getFreeSpace());
+			f.delete();	
+			i.msgb("O aquivo " +nome +" foi removido com sucesso!");			
+		}else{			
+			i.msgb("Arquivo mão encontrado!\nNão é possivél remover.");
+		}
 	} 
 	
 	
-	public void buscaArquivoExistente(final String nomeAquivo){
+	public void buscaExistente(final String nomeAquivo){
 		Path p = Paths.get(nomeAquivo);
 		i.msg("Arquivo encontrado: "+p);
 	}
-
+	
 	/**
-	 * 
 	 * @param nomeAquivo
 	 * @throws Exception
 	 */
-	public void imprimeArquivo(String nomeAquivo) throws Exception{
+	public void imprimeDados(String nomeAquivo) throws Exception{
     	try {
 			FileReader f = new FileReader(nomeAquivo);
 			BufferedReader br = new BufferedReader(f);
 			String linha = br.readLine();
-			while(linha !=null ){
-				System.out.println(linha);
+			while(linha != null ){
 				linha = br.readLine();
+				System.out.println(linha);
 			}
 			br.close();			
 		} catch (FileNotFoundException e) {
